refactor: 디렉토리 구조 변경
- components 내부의 파일들을 도메인+역할을 기준으로 한 번 더 그룹화 진행
