+++ frontend/src/components/PollMain/PollMainProfile/PollMainProfile.tsx
