+++ frontend/src/pages/GroupInitPage/GroupInitPage.tsx
@@ -4,8 +4,8 @@ import { useNavigate } from 'react-router-dom';
 
 import Logo from '../../assets/logo.svg';
 
-import GroupCreateForm from '../../components/GroupCreateForm/GroupCreateForm';
-import GroupParticipateForm from '../../components/GroupParticipateForm/GroupParticipateForm';
+import GroupCreateForm from '../../components/GroupInit/GroupCreateForm/GroupCreateForm';
+import GroupParticipateForm from '../../components/GroupInit/GroupParticipateForm/GroupParticipateForm';
 import Box from '../../components/common/Box/Box';
 import FlexContainer from '../../components/common/FlexContainer/FlexContainer';
 import { getDefaultGroup } from '../../api/group';
