+++ frontend/src/pages/PollCreatePage/PollCreatePage.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import React from 'react';
-import PollCreateForm from '../../components/PollCreateForm/PollCreateForm';
+import PollCreateForm from '../../components/PollCreate/PollCreateForm/PollCreateForm';
 
 function PollCreatePage() {
   return (
